requirements.txt
Django==2.1.4
django-elasticsearch-dsl==0.5.1
djangorestframework==3.9.0
elasticsearch==6.3.1
elasticsearch-dsl==6.1.0
ipaddress==1.0.22
psycopg2==2.7.6.1
python-dateutil==2.7.5
pytz==2018.7
six==1.12.0
urllib3==1.24.1
markdown
